make getRenderedSize async
Since React v16.6.0, requestIdleCallback is used internally for scheduling. The infinite while loop this library used for older versions causes a stack overflow as a result (since in a loop it is never idle). By making the handler async, it requires the developer get a bit more creative in specific scenarios, but prevents the overflow (and is better for the renderer anyway, since it is non-blocking).